+++ src/apiConfig.js
@@ -1,38 +1,50 @@
+const _ = require("lodash");
 const { formatDescription } = require("./common");
 
-const createApiConfig = ({ info, servers, }) => {
-  const server = (servers && servers[0]) || { url: '' }
+const createApiConfig = ({ info, servers }) => {
+  const server = (servers && servers[0]) || { url: "" };
 
-  const generic = [
-    'SecurityDataType'
-  ].filter(Boolean).join(', ')
+  const generic = _.compact([
+    {
+      name: "SecurityDataType",
+      defaultValue: "any",
+    },
+  ]);
+
+  const description = _.compact([
+    `@title ${info.title || "Api"}`,
+    info.version && `@version ${info.version}`,
+    server.url && `@baseUrl ${server.url}`,
+    _.replace(formatDescription(info.description), /\n/g, "\n * "),
+  ]);
 
   return {
     props: [
       {
-        name: 'baseUrl',
+        name: "baseUrl",
         optional: true,
-        type: 'string'
+        type: "string",
       },
       {
-        name: 'baseApiParams',
+        name: "baseApiParams",
         optional: true,
-        type: 'RequestParams'
+        type: "RequestParams",
       },
       {
-        name: 'securityWorker',
+        name: "securityWorker",
         optional: true,
-        type: '(securityData: SecurityDataType) => RequestParams'
-      }
+        type: "(securityData: SecurityDataType) => RequestParams",
+      },
     ].filter(Boolean),
-    generic: generic && `<${generic}>`,
+    generic,
     baseUrl: server.url,
     title: info.title,
     version: info.version,
-    description: formatDescription(info.description),
-  }
-}
+    description,
+    hasDescription: !!description.length,
+  };
+};
 
 module.exports = {
   createApiConfig,
-}
+};